lib: distinguish webidl interfaces with the extended property "Exposed"
