+++ lib/internal/bootstrap/web/exposed-wildcard.js
@@ -1,17 +1,21 @@
 'use strict';
 
+/**
+ * This file exposes web interfaces that is defined with the WebIDL
+ * [Exposed=*] extended attribute.
+ * See more details at https://webidl.spec.whatwg.org/#Exposed.
+ */
+
 const {
-  ObjectDefineProperty,
   globalThis,
 } = primordials;
 
 const {
-  defineOperation,
   exposeInterface,
   lazyDOMExceptionClass,
-  defineLazyProperties,
-  defineReplaceableLazyAttribute,
   exposeLazyInterfaces,
+  exposeGetterAndSetter,
+  exposeNamespace,
 } = require('internal/util');
 const config = internalBinding('config');
 
@@ -35,36 +39,17 @@ exposeGetterAndSetter(globalThis,
                         exposeInterface(globalThis, 'DOMException', value);
                       });
 
-// https://html.spec.whatwg.org/multipage/webappapis.html#windoworworkerglobalscope
-const timers = require('timers');
-defineOperation(globalThis, 'clearInterval', timers.clearInterval);
-defineOperation(globalThis, 'clearTimeout', timers.clearTimeout);
-defineOperation(globalThis, 'setInterval', timers.setInterval);
-defineOperation(globalThis, 'setTimeout', timers.setTimeout);
-
+// https://dom.spec.whatwg.org/#interface-abortcontroller
 // Lazy ones.
-exposeLazyInterfaces(globalThis, 'internal/worker/io', ['BroadcastChannel']);
 exposeLazyInterfaces(globalThis, 'internal/abort_controller', [
   'AbortController', 'AbortSignal',
 ]);
+// https://dom.spec.whatwg.org/#interface-eventtarget
 const {
   EventTarget, Event,
 } = require('internal/event_target');
 exposeInterface(globalThis, 'Event', Event);
 exposeInterface(globalThis, 'EventTarget', EventTarget);
-exposeLazyInterfaces(globalThis, 'internal/worker/io', [
-  'MessageChannel', 'MessagePort', 'MessageEvent',
-]);
-defineLazyProperties(globalThis, 'buffer', ['atob', 'btoa']);
-// https://www.w3.org/TR/FileAPI/#dfn-Blob
-exposeLazyInterfaces(globalThis, 'internal/blob', ['Blob']);
-// https://www.w3.org/TR/hr-time-2/#the-performance-attribute
-exposeLazyInterfaces(globalThis, 'perf_hooks', [
-  'Performance', 'PerformanceEntry', 'PerformanceMark', 'PerformanceMeasure',
-  'PerformanceObserver', 'PerformanceObserverEntryList', 'PerformanceResourceTiming',
-]);
-
-defineReplaceableLazyAttribute(globalThis, 'perf_hooks', ['performance']);
 
 // https://encoding.spec.whatwg.org/#textencoder
 // https://encoding.spec.whatwg.org/#textdecoder
@@ -85,27 +70,6 @@ function createGlobalConsole() {
     setConsoleExtensionInstaller(inspector.installConsoleExtensions);
   }
   return consoleFromNode;
-}
-
-// https://heycam.github.io/webidl/#es-namespaces
-function exposeNamespace(target, name, namespaceObject) {
-  ObjectDefineProperty(target, name, {
-    __proto__: null,
-    writable: true,
-    enumerable: false,
-    configurable: true,
-    value: namespaceObject,
-  });
-}
-
-function exposeGetterAndSetter(target, name, getter, setter = undefined) {
-  ObjectDefineProperty(target, name, {
-    __proto__: null,
-    enumerable: false,
-    configurable: true,
-    get: getter,
-    set: setter,
-  });
 }
 
 // Web Streams API
