+++ lib/internal/process/pre_execution.js
@@ -268,7 +268,7 @@ function setupFetch() {
   });
 }
 
-// TODO(aduh95): move this to internal/bootstrap/browser when the CLI flag is
+// TODO(aduh95): move this to internal/bootstrap/web/* when the CLI flag is
 //               removed.
 function setupWebCrypto() {
   if (process.config.variables.node_no_browser_globals ||
@@ -316,7 +316,7 @@ function setupCodeCoverage() {
   }
 }
 
-// TODO(daeyeon): move this to internal/bootstrap/browser when the CLI flag is
+// TODO(daeyeon): move this to internal/bootstrap/web/* when the CLI flag is
 //                removed.
 function setupCustomEvent() {
   if (process.config.variables.node_no_browser_globals ||
