+++ lib/internal/url.js
@@ -91,11 +91,6 @@ const {
   updateUrl,
 } = internalBinding('url');
 
-const {
-  storeDataObject,
-  revokeDataObject,
-} = internalBinding('blob');
-
 const FORWARD_SLASH = /\//g;
 
 const context = Symbol('context');
@@ -718,35 +713,6 @@ class URL {
   toJSON() {
     return this.#context.href;
   }
-
-  static createObjectURL(obj) {
-    const cryptoRandom = lazyCryptoRandom();
-    if (cryptoRandom === undefined)
-      throw new ERR_NO_CRYPTO();
-
-    const blob = lazyBlob();
-    if (!blob.isBlob(obj))
-      throw new ERR_INVALID_ARG_TYPE('obj', 'Blob', obj);
-
-    const id = cryptoRandom.randomUUID();
-
-    storeDataObject(id, obj[blob.kHandle], obj.size, obj.type);
-
-    return `blob:nodedata:${id}`;
-  }
-
-  static revokeObjectURL(url) {
-    url = `${url}`;
-    try {
-      // TODO(@anonrig): Remove this try/catch by calling `parse` directly.
-      const parsed = new URL(url);
-      const split = StringPrototypeSplit(parsed.pathname, ':');
-      if (split.length === 2)
-        revokeDataObject(split[1]);
-    } catch {
-      // If there's an error, it's ignored.
-    }
-  }
 }
 
 ObjectDefineProperties(URL.prototype, {
@@ -767,10 +733,58 @@ ObjectDefineProperties(URL.prototype, {
   toJSON: kEnumerableProperty,
 });
 
-ObjectDefineProperties(URL, {
-  createObjectURL: kEnumerableProperty,
-  revokeObjectURL: kEnumerableProperty,
-});
+function installObjectURLMethods() {
+  const {
+    storeDataObject,
+    revokeDataObject,
+  } = internalBinding('blob');
+
+  function createObjectURL(obj) {
+    const cryptoRandom = lazyCryptoRandom();
+    if (cryptoRandom === undefined)
+      throw new ERR_NO_CRYPTO();
+
+    const blob = lazyBlob();
+    if (!blob.isBlob(obj))
+      throw new ERR_INVALID_ARG_TYPE('obj', 'Blob', obj);
+
+    const id = cryptoRandom.randomUUID();
+
+    storeDataObject(id, obj[blob.kHandle], obj.size, obj.type);
+
+    return `blob:nodedata:${id}`;
+  }
+
+  function revokeObjectURL(url) {
+    url = `${url}`;
+    try {
+      // TODO(@anonrig): Remove this try/catch by calling `parse` directly.
+      const parsed = new URL(url);
+      const split = StringPrototypeSplit(parsed.pathname, ':');
+      if (split.length === 2)
+        revokeDataObject(split[1]);
+    } catch {
+      // If there's an error, it's ignored.
+    }
+  }
+
+  ObjectDefineProperties(URL, {
+    createObjectURL: {
+      __proto__: null,
+      configurable: true,
+      writable: true,
+      enumerable: true,
+      value: createObjectURL,
+    },
+    revokeObjectURL: {
+      __proto__: null,
+      configurable: true,
+      writable: true,
+      enumerable: true,
+      value: revokeObjectURL,
+    },
+  });
+}
 
 // application/x-www-form-urlencoded parser
 // Ref: https://url.spec.whatwg.org/#concept-urlencoded-parser
@@ -1251,6 +1265,7 @@ module.exports = {
   fileURLToPath,
   pathToFileURL,
   toPathIfFileURL,
+  installObjectURLMethods,
   URL,
   URLSearchParams,
   domainToASCII,
