+++ lib/internal/util.js
@@ -511,6 +511,27 @@ function exposeInterface(target, name, interfaceObject) {
   });
 }
 
+// https://heycam.github.io/webidl/#es-namespaces
+function exposeNamespace(target, name, namespaceObject) {
+  ObjectDefineProperty(target, name, {
+    __proto__: null,
+    writable: true,
+    enumerable: false,
+    configurable: true,
+    value: namespaceObject,
+  });
+}
+
+function exposeGetterAndSetter(target, name, getter, setter = undefined) {
+  ObjectDefineProperty(target, name, {
+    __proto__: null,
+    enumerable: false,
+    configurable: true,
+    get: getter,
+    set: setter,
+  });
+}
+
 function defineLazyProperties(target, id, keys, enumerable = true) {
   const descriptors = { __proto__: null };
   let mod;
@@ -750,6 +771,8 @@ module.exports = {
   emitExperimentalWarning,
   exposeInterface,
   exposeLazyInterfaces,
+  exposeNamespace,
+  exposeGetterAndSetter,
   filterDuplicateStrings,
   filterOwnProperties,
   getConstructorOf,
